Fix:change directory
