+++ tests/Feature/Controllers/ProvincieTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests\Feature\Models;
+namespace Controllers;
 
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Tests\TestCase;
